Restructure feature views, linting
