+++ rn/Teacher/ios/Teacher/Routes.swift
@@ -270,7 +270,7 @@ let router = Router(routes: HelmManager.shared.routeHandlers([
     },
 
     "/profile": { _, _, _ in
-        return CoreHostingController(SideMenu(.teacher), isDrawer: true)
+        return CoreHostingController(SideMenuView(.teacher), isDrawer: true)
     },
 
     "/profile/settings": { _, _, _ in
